requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
oracledb>=2.0.0
sqlalchemy==2.0.23
python-docx==1.1.0
openpyxl==3.1.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
pyyaml==6.0.1
